move fba load fxn
